Added delete confirmation page for deleting product for super user

+++ products/views.py
@@ -203,10 +203,17 @@ def delete_product(request, product_id):
     if not request.user.is_superuser:
         messages.error(request, 'Sorry, only store owners can do that.')
         return redirect(reverse('home'))
-    product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+    product = get_object_or_404(Product, pk=product_id)    
+    
+    if request.method == 'POST':
+        product.delete()
+        messages.success(request, 'Product deleted!')
+        return redirect(reverse('products'))
+
+    context = {
+        'product': product,
+    }
+    return render(request, 'products/delete_product.html', context)
 
 
 @login_required
